+++ AIPUBuilder/Parser/ops/release_ops.py
@@ -1123,7 +1123,7 @@ class ArmConvTranspose3DOp(BaseActivationOp, BaseConvOp, ArmOp):
         inputs = self.get_input_tensors()
         out_shape = [inputs[0].shape[0]] + \
             self.output_shape + [self.num_output]
-        out_tensor = np.random.ranf(size=out_shape).astype(np.float32)
+        out_tensor = np.random.ranf(size=out_shape).astype(inputs[0].dtype)
         self.set_out_tensor(out_tensor)
